Small code and language fixes

+++ core/src/commonMain/kotlin/maryk/core/processors/datastore/readStorageToChanges.kt
@@ -477,7 +477,7 @@ private fun <P : PropertyDefinitions> readQualifierOfType(
                 qualifierLength,
                 offset,
                 readValueFromStorage,
-                index.toUInt(),
+                index,
                 reference,
                 select,
                 addToCache,
